requirements.txt
Django==4.0.5
python-telegram-bot==13.12
python_dotenv==0.20.0
